feat: refactors directory services for modularity
