+++ apps/internal-cli/src/commands/directory/delete.subcommand.ts
@@ -4,7 +4,7 @@ import chalk from 'chalk';
 import ora from 'ora';
 import inquirer from 'inquirer';
 import { DirectoryRepository, UserRepository } from '@packages/agent/database';
-import { AgentService } from '@packages/agent/services';
+import { DirectoryLifecycleService } from '@packages/agent/services';
 import { DirectoryPromptService } from './directory-prompt.service';
 import { ConfigCheckService } from './config-check.service';
 import { handleCliError } from './error';
@@ -20,7 +20,7 @@ export class DeleteSubCommand extends CommandRunner {
         private readonly directoryRepository: DirectoryRepository,
         private readonly directoryPrompt: DirectoryPromptService,
         private readonly configCheck: ConfigCheckService,
-        private readonly agentService: AgentService,
+        private readonly directoryLifecycleService: DirectoryLifecycleService,
         private readonly userRepository: UserRepository,
     ) {
         super();
@@ -126,7 +126,7 @@ export class DeleteSubCommand extends CommandRunner {
                 const user = await this.userRepository.createOrGetLocalUser();
 
                 // Call the agent service method directly
-                const result = await this.agentService.deleteDirectory(
+                const result = await this.directoryLifecycleService.deleteDirectory(
                     directory.id,
                     deleteOptions,
                     user,
