+++ packages/agent/src/services/index.ts
@@ -1,3 +1,6 @@
-export * from './agent.service';
 export * from './agent.module';
 export * from './directory-detail.service';
+export * from './directory-ownership.service';
+export * from './directory-query.service';
+export * from './directory-lifecycle.service';
+export * from './directory-generation.service';
